Added servo control for python scripts

+++ Arduino/LightMapper/LightMapper.ino
@@ -165,9 +165,9 @@ void stepMotorsIfTime() {
  * degrees_byte: The servo control byte
  */
 void setServoPos(uint8_t degrees_byte) {
-  int8_t servo_degrees = static_cast<int8_t>(degrees_byte);
-  int8_t servo_center = (SERVO_MAX_DEGREES + SERVO_MIN_DEGREES) / 2;
-  int8_t translated_degrees = servo_center + servo_degrees;
+  int8_t servo_degrees = static_cast<int16_t>(degrees_byte);
+  int16_t servo_center = (SERVO_MAX_DEGREES + SERVO_MIN_DEGREES) / 2;
+  int16_t translated_degrees = static_cast<int16_t>(servo_center + servo_degrees);
 
   // Set the motor position if the position is valid
   if ((translated_degrees < SERVO_MIN_DEGREES) || (translated_degrees > SERVO_MAX_DEGREES)) {
